+++ src/ROSA/foreign_potato_master.py
@@ -38,7 +38,7 @@ as well
 '''
 
 # note: view associated GitHub info as well
-__version__ = 'v0.6.0'  
+__version__ = 'v0.7.0'  
 __author__ = 'Cornelius-Figgle'
 __email__ = 'max@fullimage.net'
 __maintainer__ = 'Cornelius-Figgle'
